codestyle. Поправлены некоторые моменты по код стайлу, согласно ревью. feat. добавлена обработка остальных исключений. refactor. переделана логика, теперь алгоритме отображения общих друзей используется stream.

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -13,7 +13,7 @@ import java.util.Set;
 @Valid
 public class Film {
     @JsonIgnore
-    private Set<Integer> likes = new HashSet<>();
+    private Set<Long> likes = new HashSet<>();
     @JsonIgnore
     private int likesCount;
     private long id;
